Files_Versioning-Backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
grpcio==1.51.3
grpcio-tools==1.51.3
gunicorn==20.1.0
inflection==0.5.1
jsonify==0.5
packaging==23.1
protobuf==4.22.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
